Add back the deprecated function create_database_backup
This function was removed when it was replaced with the renamed
version create_db_backup.

That created what would be a breaking change in the API, for users
depending on the old name.

This commit re-introduces the function, with a deprecation warning to
alert users that it will be removed in future versions.

This will make the next release non-breaking.